third_party/WebKit/LayoutTests/fast/dom/custom/imports-block-upgrade-in-master.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
var upgradeIds = [];
</script>
<x-a id='a1'></x-a>
<link rel="import" href="resources/import-block-upgrade-in-master-child.html">
<body>
<x-a id='a2'></x-a>
<script>
test(function () {
    assert_array_equals(window.upgradeIds, ['a1', 'a2'], 'Elements in the master should be upgraded by a definition in an import.');
}, 'upgrade in the master');
</script>
</body>
